requirements.txt
requests==2.32.3
pandas==2.2.2
numpy==2.0.1
ta>=0.11
APScheduler>=3.10
pytz>=2024.1
python-dotenv==1.0.1
requests==2.32.3
python-telegram-bot==21.4
python-binance==1.0.19
python-telegram-bot[rate-limiter]==20.7
